requirements.txt
Django==4.1.3
djangorestframework==3.14.0
openpyxl==3.0.10
